requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
djangorestframework==3.14.0
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
